Create a CompactRestaurantInfo component
Transfer code from map-callout component to it and Import it and use it.
